.bash_aliases: upwards-recursive `sba`
Shell functions are kind of like aliases, right? In any case it's
certainly not more of sin than thinking of this endeavor as a "dotfiles"
repository, when not every important portable system configuration file
starts with a dot.

Thanks to http://superuser.com/a/724717 for the tip about how to search
upwards (aesthetically, I don't like the way in which it actually
changes directories while you're too slow to look, but it works and it's
pretty readable).