update git-wordcount
I, um, don't actually know the arguments to `git diff` (I think you
can specify a range of commits with an ellipsis??) and have always
been using this either with no-args, or with a single SHA. Let's
generalize to multiple SHAs in the obvious way. (Maybe I should have
read the `git-diff` man page first if that already does multiple SHAs
in the obvious way? Um.)